feat: Adicionando funções para edição de nome e email do usuário

+++ src/controllers/user.controller.js
@@ -61,12 +61,22 @@ export class UserController {
   }
 
   static async delete(req, res) {
-    const { user_id } = req.body;
-    UserContentRepository.deleteUserContentByUserId(user_id);
-    UserPathRepository.deleteUserPathByUserId(user_id);
-    UserRepository.deleteUserById(user_id);
+    const { user_id, password } = req.body;
+    const user = await UserRepository.getUserById(user_id);
 
-    return res.status(200).json({ msg: "Usuário excluído com sucesso." });
+    if (!user) {
+      res.status(404).json({ msg: "Usuário não existe" });
+    }
+
+    if (password === user.password) {
+      UserContentRepository.deleteUserContentByUserId(user_id);
+      UserPathRepository.deleteUserPathByUserId(user_id);
+      UserRepository.deleteUserById(user_id);
+
+      return res.status(200).json({ msg: "Usuário excluído com sucesso." });
+    }
+
+    return res.status(404).json({ msg: "Senha incorreta." });
   }
 
   static async changePassword(req, res) {
@@ -114,6 +124,7 @@ export class UserController {
       user_id: user.user_id,
       name: user.name,
       email: user.email,
+      password: user.password,
       is_admin: user.is_admin,
     });
   }
